refactor: Rename Class And Do Clean Code
